messages-service/requirements.txt
flask
flask-restful
pika
python-consul
